// packages/astro/dist/assets/types.d.ts
export type ImageQuality = ImageQualityPreset | number;
export type ImageInputFormat = (typeof VALID_INPUT_FORMATS)[number];
export type ImageOutputFormat = (typeof VALID_OUTPUT_FORMATS)[number] | (string & {});
export type ImageLayout = 'constrained' | 'fixed' | 'full-width' | 'none';
export type ImageFit = 'fill' | 'contain' | 'cover' | 'none' | 'scale-down' | (string & {});
// ...
declare const isESMImport: unique symbol;
export type OmitBrand<T> = Omit<T, typeof isESMImport>;
/**
 * Type returned by ESM imports of images
 */
export type ImageMetadata = {
	src: string;
	width: number;
	height: number;
	format: ImageInputFormat;
	orientation?: number;
	[isESMImport]?: true;
};
export declare function isImageMetadata(src: any): src is ImageMetadata;
/**
 * A yet to be completed with an url `SrcSetValue`. Other hooks will only see a resolved value, where the URL of the image has been added.
 */
export type UnresolvedSrcSetValue = {
	transform: ImageTransform;
	descriptor?: string;
	attributes?: Record<string, any>;
};
export type SrcSetValue = UnresolvedSrcSetValue & {
	url: string;
};
/**
 * A yet to be resolved image transform. Used by `getImage`
 */
export type UnresolvedImageTransform = Simplify<
	OmitPreservingIndexSignature<ImageTransform, 'src'> & {
		src:
			| ImageMetadata
			| string
			| Promise<{
					default: ImageMetadata;
			  }>;
		inferSize?: boolean;
	}
> & {
	[isESMImport]?: never;
};
/**
 * Options accepted by the image transformation service.
 */
export type ImageTransform = {
	src: ImageMetadata | string;
	width?: number | undefined;
	widths?: number[] | undefined;
	densities?: (number | `${number}x`)[] | undefined;
	height?: number | undefined;
	quality?: ImageQuality | undefined;
	format?: ImageOutputFormat | undefined;
	fit?: ImageFit | undefined;
	position?: string | undefined;
	background?: string | undefined;
	[key: string]: any;
} & Astro.CustomImageProps;
export interface GetImageResult {
	rawOptions: ImageTransform;
	options: ImageTransform;
	src: string;
	srcSet: {
		values: SrcSetValue[];
		attribute: string;
	};
	attributes: Record<string, any>;
}